Rewrite problem statement to use an UNBIASED error function.
Talking it over with Georgia revealed that I was confused (or at least that I'd made a poor
choice of toy formulation). Namely, if your estimator of the box value is unbiased, you now
have some reason to believe it (in the sense of assigning an expected value equal to the
label). This makes the curse more visible; when you've selected the highest label from among
multiple boxes, the label becomes an overestimate of the expected value.

Well, so I think, anyways. Further commits will actually model this.